42 is a double

+++ jni/demo/jni/fourtyTwo.c
@@ -1,5 +1,5 @@
 // Return the answer to life
 // Return "the meaning of life, the universe, and everything"
-int meaningOfLife() {
-  return 42;
+double meaningOfLife() {
+  return 42.0;
 }
